+++ guillotina/api/__init__.py
@@ -7,9 +7,11 @@ from . import content  # noqa
 from . import dynamic  # noqa
 from . import files  # noqa
 from . import login  # noqa
+from . import metadata  # noqa
 from . import registry  # noqa
 from . import search  # noqa
 from . import storage  # noqa
+from . import suggestion  # noqa
 from . import types  # noqa
 from . import user  # noqa
 from . import vocabularies  # noqa